Confirmed functionality of Nodemailer to gmail addresses. Relocated script to emails.js inside of controllers/api
