+++ controllers/api/emailer.js
@@ -1,0 +1,34 @@
+"use strict";
+const nodemailer = require("nodemailer");
+require('dotenv').config();
+
+// async..await is not allowed in global scope, must use a wrapper
+async function main() {
+
+  // create reusable transporter object using the default SMTP transport
+  let transporter = nodemailer.createTransport({
+    service: 'gmail',
+    auth: {
+      user: process.env.EMAIL_USER, // generated ethereal user
+      pass: process.env.EMAIL_PASSWORD, // generated ethereal password
+    },
+  });
+
+  // send mail with defined transport object this object shouldn't change, other than the html/text fields
+  let info = await transporter.sendMail({
+    from: '"Reminder Services ⛑️ - Do Not Reply" <mymedreminder@gmail.com', // sender address
+    to: "joe.d.malone@gmail.com", // list of receivers
+    subject: "Hello ✔", // Subject line
+    // text: "Hello world?", // plain text body
+    html: "<b>Hello world?</b>", // html body
+  });
+
+  console.log("Message sent: %s", info.messageId);
+  // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
+
+  // Preview only available when sending through an Ethereal account
+  console.log("Preview URL: %s", nodemailer.getTestMessageUrl(info));
+  // Preview URL: https://ethereal.email/message/WaQKMgKddxQDoou...
+}
+
+main().catch(console.error);
