--- controllers/api/user-routes.js
@@ -1,54 +1,0 @@
-"use strict";
-const nodemailer = require("nodemailer");
-require('dotenv').config();
-
-// const transporterDefaults = {/*create an object for persistent values, such as 'From' Email address, to be included with each sent message*/};
-// let transporter = nodemailer.createTransport(options[, defaults])
-
-
-
-// async..await is not allowed in global scope, must use a wrapper
-async function main() {
-  // Generate test SMTP service account from ethereal.email
-  // Only needed if you don't have a real mail account for testing
-  let testAccount = await nodemailer.createTestAccount();
-
-  // create reusable transporter object using the default SMTP transport
-  let transporter = nodemailer.createTransport({
-    host: "smtp.gmail.com",
-    port: 587,
-    secure: false, // true for 465, false for other ports
-    auth: {
-      user: process.env.EMAIL_USER, // generated ethereal user
-      pass: process.env.EMAIL_PASSWORD, // generated ethereal password
-    },
-  });
-
-  // send mail with defined transport object
-  let info = await transporter.sendMail({
-    from: '"Fred Foo 👻" <foo@example.com>', // sender address
-    to: "joe.d.malone@gmail.com", // list of receivers
-    subject: "Hello ✔", // Subject line
-    text: "Hello world?", // plain text body
-    // html: "<b>Hello world?</b>", // html body
-  });
-
-  console.log("Message sent: %s", info.messageId);
-  // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
-
-  // Preview only available when sending through an Ethereal account
-  console.log("Preview URL: %s", nodemailer.getTestMessageUrl(info));
-  // Preview URL: https://ethereal.email/message/WaQKMgKddxQDoou...
-}
-
-main().catch(console.error);
-
-
-
-// var message = {
-//     from: "sender@server.com",
-//     to: "receiver@sender.com",
-//     subject: "Medication Reminder",
-//     text: "Plaintext version of the message",
-//     html: "<p>HTML version of the message</p>"
-//   };
